requirements.txt
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Dropzone==1.5.4
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
Flask-Avatars==0.2.2
Flask-Mail==0.9.1
flask-whooshee==0.7.0
mistune==0.8.4
Pillow==6.0.0
pypinyin==0.35.3
python-dotenv==0.10.3
beautifulsoup4==4.8.2
PyJWT==1.7.1
elasticsearch==7.1.0

#requirements for Heroku
#psycopg2==2.7.3.1
#gunicorn==19.7.1